Remove IdentifierContainerInterface
